pyproject.toml
[project]
name = "my-spec-cli"
version = "0.1.0"
description = "A simplified SDD CLI inspired by Spec Kit"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.9.0",         # 命令行界面
    "rich>=13.0.0",         # 漂亮的终端输出
    "httpx>=0.25.0",        # 网络请求 (比requests更现代化)
    "readchar>=4.0.0",      # 键盘输入处理
    "truststore>=0.8.0",    # 系统SSL证书支持
    "sentence-transformers>=3.0.0",
    "faiss-cpu>=1.8.0",
    "numpy>=2.0.0",
    "rank-bm25>=0.2.2",
    "jieba>=0.42.1"
]

[project.scripts]
myspec = "my_cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[tool.hatch.build.targets.wheel]
packages = ["src/my_cli"]